+++ apps/user/src/components/common/index.ts
@@ -1,4 +1,5 @@
 export { default as Footer } from './Footer/Footer';
 export { default as Header } from './Header/Header';
 export { default as Loader } from './Loader/Loader';
-export { default as AuthWrapper } from './AuthWrapper/AuthWrapper';
+export { default as AuthWrapper } from './Wrappers/AuthWrapper/AuthWrapper';
+export { default as FormWrapper } from './Wrappers/FormWrapper/FormWrapper';
